Move MetadataStore into restate-core
