+++ crates/metadata-store/src/local/mod.rs
@@ -14,12 +14,12 @@ mod store;
 mod service;
 
 pub use options::Options;
+use restate_core::metadata_store::MetadataStoreClient;
 use restate_types::net::AdvertisedAddress;
 pub use service::LocalMetadataStoreService;
 pub use store::BuildError;
 
 use crate::local::grpc::client::LocalMetadataStoreClient;
-use crate::MetadataStoreClient;
 
 /// Creates a [`MetadataStoreClient`] for the [`LocalMetadataStoreService`].
 pub fn create_client(advertised_address: AdvertisedAddress) -> MetadataStoreClient {
